ABI-1: Breakdown Cart to include CartContents component
Cart is responsible for fetching data and CartContents is responsible for working with that data.

Remove unnecessary DOM element nesting in Cart.js
